finished login, register, userprofile implementation with backend
