+++ package.json
@@ -10,6 +10,7 @@
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
     "react-icons": "^5.5.0",
+    "react-router-dom": "^7.3.0",
     "react-scripts": "^5.0.1",
     "web-vitals": "^2.1.4"
   },
