+++ src/App.js
@@ -1,11 +1,20 @@
-import LoginRegister from './Components/LoginRegister/LoginRegister';
+import React from 'react';
+import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
+import Login from './Components/Login';
+import Register from './Components/Register';
+import UserProfile from './Components/UserProfile';
 
-function App() {
-  return (
-    <div>
-      <LoginRegister/>
-    </div>
-  );
-}
+const App = () => {
+    return (
+        <Router>
+            <Routes>
+                <Route path="/login" element={<Login />} />
+                <Route path="/register" element={<Register />} />
+                <Route path="/userprofile" element={<UserProfile />} />
+                <Route path="/" element={<Login />} /> {/* Default route */}
+            </Routes>
+        </Router>
+    );
+};
 
 export default App;
